refactor: modularize main scripts
